+++ SlidingBlocksGame/Assets/_Project/Scripts/Gameplay/Features/CommonFeature/Components/CellPosition.cs
@@ -7,7 +7,7 @@ namespace _Project.Scripts.Gameplay.Features.CommonFeature.Components
     [Serializable]
     public struct CellPosition : IEcsComponent
     {
-        public float2 Value;
+        public int2 Value;
 
         [Serializable]
         public sealed class Wrapper : ComponentTemplate<CellPosition>
